VIM-4120 Display vim error messages in output panel
