+++ src/main/java/com/maddyhome/idea/vim/group/KeyGroup.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2003-2023 The IdeaVim authors
+ * Copyright 2003-2026 The IdeaVim authors
  *
  * Use of this source code is governed by an MIT-style
  * license that can be found in the LICENSE.txt file or at
@@ -388,7 +388,7 @@ public class KeyGroup extends VimKeyGroupBase implements PersistentStateComponen
 
     VimOutputPanel outputPanel = injector.getOutputPanel()
       .getOrCreate(editor, injector.getExecutionContextManager().getEditorExecutionContext(editor));
-    outputPanel.addText(builder.toString(), true);
+    outputPanel.addText(builder.toString(), true, null);
     outputPanel.show();
     return true;
   }
