+++ src/main/resources/META-INF/plugin.xml
@@ -1,5 +1,5 @@
 <!--
-  ~ Copyright 2003-2023 The IdeaVim authors
+  ~ Copyright 2003-2026 The IdeaVim authors
   ~
   ~ Use of this source code is governed by an MIT-style
   ~ license that can be found in the LICENSE.txt file or at
@@ -63,7 +63,7 @@
               topic="com.intellij.ide.plugins.DynamicPluginListener"/>
     <listener class="com.maddyhome.idea.vim.ui.ex.ExEntryPanel$LafListener"
               topic="com.intellij.ide.ui.LafManagerListener"/>
-    <listener class="com.maddyhome.idea.vim.ui.ExOutputPanel$LafListener"
+    <listener class="com.maddyhome.idea.vim.ui.OutputPanel$LafListener"
               topic="com.intellij.ide.ui.LafManagerListener"/>
     <listener class="com.maddyhome.idea.vim.extension.highlightedyank.HighlightColorResetter"
               topic="com.intellij.ide.ui.LafManagerListener"/>
