+++ src/test/java/org/jetbrains/plugins/ideavim/ex/implementation/commands/BufferListCommandTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2003-2023 The IdeaVim authors
+ * Copyright 2003-2026 The IdeaVim authors
  *
  * Use of this source code is governed by an MIT-style
  * license that can be found in the LICENSE.txt file or at
@@ -9,7 +9,7 @@
 package org.jetbrains.plugins.ideavim.ex.implementation.commands
 
 import com.maddyhome.idea.vim.api.injector
-import com.maddyhome.idea.vim.ex.ExOutputModel.Companion.getInstance
+import com.maddyhome.idea.vim.ui.OutputModel.Companion.getInstance
 import org.jetbrains.plugins.ideavim.VimTestCase
 import org.junit.jupiter.api.Test
