+++ src/test/java/org/jetbrains/plugins/ideavim/ex/implementation/commands/ExecuteCommandTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2003-2023 The IdeaVim authors
+ * Copyright 2003-2026 The IdeaVim authors
  *
  * Use of this source code is governed by an MIT-style
  * license that can be found in the LICENSE.txt file or at
@@ -24,7 +24,7 @@ class ExecuteCommandTest : VimTestCase() {
   fun `test execute with range`() {
     configureByText("\n")
     typeText(commandToKeys("1,2execute 'echo 42'"))
-    assertNoOutput()
+    assertOutput("E481: No range allowed")
     assertPluginError(true)
   }
