+++ src/test/java/org/jetbrains/plugins/ideavim/ex/implementation/commands/HistoryCommandTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2003-2025 The IdeaVim authors
+ * Copyright 2003-2026 The IdeaVim authors
  *
  * Use of this source code is governed by an MIT-style
  * license that can be found in the LICENSE.txt file or at
@@ -8,7 +8,7 @@
 
 package org.jetbrains.plugins.ideavim.ex.implementation.commands
 
-import com.maddyhome.idea.vim.ex.ExOutputModel
+import com.maddyhome.idea.vim.ui.OutputModel
 import org.jetbrains.plugins.ideavim.VimBehaviorDiffers
 import org.jetbrains.plugins.ideavim.VimTestCase
 import org.junit.jupiter.api.BeforeEach
@@ -36,7 +36,7 @@ class HistoryCommandTest : VimTestCase() {
   @Test
   fun `test history lists all entries in cmd history by default`() {
     repeat(10) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history cmd",
       """
         |      #  cmd history
@@ -67,8 +67,7 @@ class HistoryCommandTest : VimTestCase() {
   fun `test history with 'history' option set to 0 shows nothing`() {
     enterCommand("set history=0")
     enterCommand("history")
-    assertNoOutput()
-    assertPluginError(false)
+    assertPluginError(true)
     assertPluginErrorMessage("'history' option is zero")
   }
 
@@ -98,7 +97,7 @@ class HistoryCommandTest : VimTestCase() {
   fun `test history adds indicator to current entry`() {
     repeat(5) { i -> enterSearch("foo${i + 1}") }
     repeat(5) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history all",
       """
         |      #  cmd history
@@ -123,7 +122,7 @@ class HistoryCommandTest : VimTestCase() {
   @Test
   fun `test history does not show indicator if not including current entry`() {
     repeat(10) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history : 1,5",
       """
         |      #  cmd history
@@ -139,7 +138,7 @@ class HistoryCommandTest : VimTestCase() {
   @Test
   fun `test history with no name and first number lists single entry from command history`() {
     repeat(10) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history 3",
       """
         |      #  cmd history
@@ -150,7 +149,7 @@ class HistoryCommandTest : VimTestCase() {
   @Test
   fun `test history with no name and two numbers lists command history range`() {
     repeat(10) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history 3, 6",
       """
         |      #  cmd history
@@ -174,7 +173,7 @@ class HistoryCommandTest : VimTestCase() {
   @Test
   fun `test history with colon and first number lists single entry from command history`() {
     repeat(10) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history : 3",
       """
         |      #  cmd history
@@ -185,7 +184,7 @@ class HistoryCommandTest : VimTestCase() {
   @Test
   fun `test history with colon and no space before first number lists single entry from command history`() {
     repeat(10) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history :3",
       """
         |      #  cmd history
@@ -196,7 +195,7 @@ class HistoryCommandTest : VimTestCase() {
   @Test
   fun `test history with colon and two numbers lists command history range`() {
     repeat(10) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history : 3, 6",
       """
         |      #  cmd history
@@ -254,7 +253,7 @@ class HistoryCommandTest : VimTestCase() {
   @Test
   fun `test history cmd with first number lists single entry from command history`() {
     repeat(10) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history cmd 3",
       """
         |      #  cmd history
@@ -265,7 +264,7 @@ class HistoryCommandTest : VimTestCase() {
   @Test
   fun `test history cmd with no space before first number lists single entry from command history`() {
     repeat(10) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history cmd3",
       """
         |      #  cmd history
@@ -276,7 +275,7 @@ class HistoryCommandTest : VimTestCase() {
   @Test
   fun `test history cmd with two numbers lists command history range`() {
     repeat(10) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history cmd 3, 6",
       """
         |      #  cmd history
@@ -290,14 +289,14 @@ class HistoryCommandTest : VimTestCase() {
   @Test
   fun `test history cmd with two numbers incorrectly ordered lists nothing from command history range`() {
     repeat(10) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history cmd 6,3", "      #  cmd history")
   }
 
   @Test
   fun `test history cmd with number that is no longer used`() {
     repeat(10) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     // This will make "echo 1" the last used entry, remove it from position 1 and add it at position 11
     typeText(":<Up><Up><Up><Up><Up><Up><Up><Up><Up><Up><Esc>")
     assertCommandOutput("history cmd 1", "      #  cmd history")
@@ -306,7 +305,7 @@ class HistoryCommandTest : VimTestCase() {
   @Test
   fun `test history cmd with range starting from number that is no longer used`() {
     repeat(10) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     // This will make "echo 1" the last used entry, remove it from position 1 and add it at position 11
     typeText(":<Up><Up><Up><Up><Up><Up><Up><Up><Up><Up><Esc>")
     assertCommandOutput("history cmd 1,10",
@@ -327,7 +326,7 @@ class HistoryCommandTest : VimTestCase() {
   @Test
   fun `test history cmd -1 shows last entry`() {
     repeat(10) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history cmd -1",
       """
         |      #  cmd history
@@ -338,7 +337,7 @@ class HistoryCommandTest : VimTestCase() {
   @Test
   fun `test history cmd with negative number shows list of entries relative to last entry`() {
     repeat(10) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history cmd -4,-1",
       """
         |      #  cmd history
@@ -352,14 +351,14 @@ class HistoryCommandTest : VimTestCase() {
   @Test
   fun `test history cmd with two negative numbers incorrectly ordered lists nothing from command history range`() {
     repeat(10) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history cmd -1,-4", "      #  cmd history")
   }
 
   @Test
   fun `test history with positive start number and negative last number`() {
     repeat(10) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history cmd 4,-3",
       """
         |      #  cmd history
@@ -375,7 +374,7 @@ class HistoryCommandTest : VimTestCase() {
   @Test
   fun `test history with negative start number and positive last number`() {
     repeat(10) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history cmd -8,8",
       """
         |      #  cmd history
@@ -460,7 +459,7 @@ class HistoryCommandTest : VimTestCase() {
   @Test
   fun `test history search with first number lists single entry from saerch history`() {
     repeat(10) { i -> enterSearch("foo${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history search 3",
       """
         |      #  search history
@@ -471,7 +470,7 @@ class HistoryCommandTest : VimTestCase() {
   @Test
   fun `test history search with no space before first number lists single entry from search history`() {
     repeat(10) { i -> enterSearch("foo${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history search3",
       """
         |      #  search history
@@ -482,7 +481,7 @@ class HistoryCommandTest : VimTestCase() {
   @Test
   fun `test history search with two numbers lists search history range`() {
     repeat(10) { i -> enterSearch("foo${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history search 3, 6",
       """
         |      #  search history
@@ -537,7 +536,7 @@ class HistoryCommandTest : VimTestCase() {
   fun `test history all includes history entries`() {
     repeat(5) { i -> enterSearch("foo${i + 1}") }
     repeat(5) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history all",
       """
         |      #  cmd history
@@ -563,7 +562,7 @@ class HistoryCommandTest : VimTestCase() {
   fun `test history all applies first number to all history entries`() {
     repeat(5) { i -> enterSearch("foo${i + 1}") }
     repeat(5) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history all 3",
       """
         |      #  cmd history
@@ -580,7 +579,7 @@ class HistoryCommandTest : VimTestCase() {
   fun `test history all applies range to all history entries`() {
     repeat(5) { i -> enterSearch("foo${i + 1}") }
     repeat(5) { i -> enterCommand("echo ${i + 1}") }
-    ExOutputModel.getInstance(fixture.editor).clear()
+    OutputModel.getInstance(fixture.editor).clear()
     assertCommandOutput("history all 2,4",
       """
         |      #  cmd history
