+++ src/test/java/org/jetbrains/plugins/ideavim/ex/implementation/statements/TryCatchTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2003-2023 The IdeaVim authors
+ * Copyright 2003-2026 The IdeaVim authors
  *
  * Use of this source code is governed by an MIT-style
  * license that can be found in the LICENSE.txt file or at
@@ -155,7 +155,7 @@ class TryCatchTest : VimTestCase() {
     )
     assertPluginError(true)
     assertPluginErrorMessage("my exception")
-    assertOutput("finally block")
+    assertOutput("finally block\nmy exception")
   }
 
   @Test
