+++ src/testFixtures/kotlin/org/jetbrains/plugins/ideavim/VimTestCase.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2003-2024 The IdeaVim authors
+ * Copyright 2003-2026 The IdeaVim authors
  *
  * Use of this source code is governed by an MIT-style
  * license that can be found in the LICENSE.txt file or at
@@ -62,7 +62,6 @@ import com.maddyhome.idea.vim.api.setToggleOption
 import com.maddyhome.idea.vim.api.visualLineToBufferLine
 import com.maddyhome.idea.vim.command.MappingMode
 import com.maddyhome.idea.vim.ex.ExException
-import com.maddyhome.idea.vim.ex.ExOutputModel
 import com.maddyhome.idea.vim.group.EffectiveIjOptions
 import com.maddyhome.idea.vim.group.GlobalIjOptions
 import com.maddyhome.idea.vim.group.IjOptions
@@ -84,6 +83,7 @@ import com.maddyhome.idea.vim.options.helpers.GuiCursorOptionHelper
 import com.maddyhome.idea.vim.options.helpers.GuiCursorType
 import com.maddyhome.idea.vim.state.mode.Mode
 import com.maddyhome.idea.vim.state.mode.inBlockSelection
+import com.maddyhome.idea.vim.ui.OutputModel
 import com.maddyhome.idea.vim.ui.ex.ExEntryPanel
 import com.maddyhome.idea.vim.vimscript.model.CommandLineVimLContext
 import com.maddyhome.idea.vim.vimscript.model.datatypes.VimFuncref
@@ -553,7 +553,7 @@ abstract class VimTestCase(private val defaultEditorText: String? = null) {
     return NeovimTesting.vimMode()
   }
 
-  fun register(char: String): String? {
+  fun register(char: String): String {
     return NeovimTesting.getMark(char)
   }
 
@@ -731,7 +731,7 @@ abstract class VimTestCase(private val defaultEditorText: String? = null) {
   }
 
   fun assertOutput(expected: String, clear: Boolean = true) {
-    val actual = injector.outputPanel.getCurrentOutputPanel()?.text
+    val actual = injector.outputPanel.getCurrentOutputPanel()?.text?.removeSuffix("\n")
     if (actual == null) {
       // If there's no output, there's a good chance we've got an error
       val message = "No Ex output" + if (injector.messages.isError()) {
@@ -747,12 +747,12 @@ abstract class VimTestCase(private val defaultEditorText: String? = null) {
       // Ex output is not cleared until the output pane is activated again (we need it to persist so that :print will
       // work when called multiple times from :global). When testing, if the previous action fails before it can
       // activate the output pane, we'll be looking at stale results.
-      ExOutputModel.getInstance(fixture.editor).clear()
+      injector.outputPanel.getCurrentOutputPanel()?.clearText()
     }
   }
 
   fun assertNoOutput() {
-    val actual = ExOutputModel.getInstance(fixture.editor).text
+    val actual = OutputModel.getInstance(fixture.editor).text
     assertEquals("", actual)
   }
