+++ vim-engine/src/main/kotlin/com/maddyhome/idea/vim/api/MessageType.kt
@@ -1,0 +1,16 @@
+/*
+ * Copyright 2003-2026 The IdeaVim authors
+ *
+ * Use of this source code is governed by an MIT-style
+ * license that can be found in the LICENSE.txt file or at
+ * https://opensource.org/licenses/MIT.
+ */
+
+package com.maddyhome.idea.vim.api
+
+import java.awt.Color
+
+enum class MessageType(val color: Color?) {
+  STANDARD(null),
+  ERROR(Color.RED)
+}
