+++ vim-engine/src/main/kotlin/com/maddyhome/idea/vim/api/VimFileBase.kt
@@ -22,7 +22,7 @@ abstract class VimFileBase : VimFile {
     val offset = editor.currentCaret().offset
     val ch = editor.text()[offset]
 
-    injector.messages.showErrorMessage(editor, toHexString(ch.code.toLong()))
+    injector.messages.showMessage(editor, toHexString(ch.code.toLong()))
   }
 
   override fun displayLocationInfo(editor: VimEditor) {
@@ -90,7 +90,7 @@ abstract class VimFileBase : VimFile {
       }
     }
 
-    injector.messages.showErrorMessage(editor, msg)
+    injector.messages.showMessage(editor, msg)
   }
 }
