+++ vim-engine/src/main/kotlin/com/maddyhome/idea/vim/common/CommandAlias.kt
@@ -36,7 +36,8 @@ sealed class CommandAlias(
       var compiledCommand = this.command
       val cleanedInput = input.trim().removePrefix(name).trim()
       if (minimumNumberOfArguments > 0 && cleanedInput.isEmpty()) {
-        injector.messages.showErrorMessage(editor = null, injector.messages.message("E471"))
+        injector.editorGroup.getFocusedEditor()
+          ?.let { injector.messages.showErrorMessage(it, injector.messages.message("E471")) }
         injector.messages.indicateError()
         return GoalCommand.Ex.EMPTY
       }
