+++ vim-engine/src/main/kotlin/com/maddyhome/idea/vim/vimscript/model/commands/GlobalCommand.kt
@@ -64,7 +64,7 @@ data class GlobalCommand(val range: Range, val modifier: CommandModifier, val ar
     // When nesting the command works on one line.  This allows for
     // ":g/found/v/notfound/command".
     if (globalBusy && (range.startLine != 0 || range.endLine != editor.lineCount() - 1)) {
-      messages.showErrorMessage(null, messages.message("E147"))
+      messages.showErrorMessage(editor, messages.message("E147"))
       messages.indicateError()
       return false
     }
@@ -120,16 +120,16 @@ data class GlobalCommand(val range: Range, val modifier: CommandModifier, val ar
       }
 
       if (gotInt) {
-        messages.showErrorMessage(null, messages.message("command.global.interrupted"))
+        messages.showErrorMessage(editor, messages.message("command.global.interrupted"))
       } else if (marks.isEmpty()) {
         if (invert) {
           messages.showErrorMessage(
-            null,
+            editor,
             messages.message("command.global.pattern.found.in.every.line", globalCommandArguments.pattern.toString())
           )
         } else {
           messages.showErrorMessage(
-            null,
+            editor,
             messages.message("command.global.pattern.not.found", globalCommandArguments.pattern.toString())
           )
         }
